optimize catalog filters and pagination
